transform_into_bigquery/requirements.txt
pandas==2.1.0
google.cloud==0.34.0
google-cloud-storage==2.10.0
google-cloud-bigquery==3.11.4